adding checks in removal script

+++ scripts/worker/remove_user.sh
@@ -3,8 +3,15 @@
 # - user login
 args=("$@")
 if [ $# -eq 1 ]; then
-    sudo userdel -f ${args[0]}
-    sudo rm -R /home/${args[0]}
+
+    if [ `egrep -i "^$username" /etc/passwd > /dev/null` ]; then
+        sudo userdel -f ${args[0]}
+    fi
+
+    if [ -d "/home/${args[0]}" ]; then
+        sudo rm -R /home/${args[0]}
+    fi    
+    
     echo 'ok'
 else
     echo 'error'
